+++ edx_django_utils/__init__.py
@@ -2,7 +2,7 @@
 EdX utilities for Django Application development..
 """
 
-__version__ = "5.14.1"
+__version__ = "5.14.2"
 
 default_app_config = (
     "edx_django_utils.apps.EdxDjangoUtilsConfig"